Remove FrameObject subclass from `stbt/__init__.py`
The subclass just existed to avoid having _stbt have access to
`stbt._dut`.  By following the pattern established in fdca78fdf39 we
do:

    if frame is None:
        import stbt
        frame = stbt.get_frame()

The local import breaks the circular import dependency.

I've moved it into it's own file under `_stbt` too.
