+++ Makefile
@@ -62,6 +62,7 @@ INSTALL_PYLIB_FILES = \
     _stbt/control.py \
     _stbt/core.py \
     _stbt/cv2_compat.py \
+    _stbt/frameobject.py \
     _stbt/gst_hacks.py \
     _stbt/gst_utils.py \
     _stbt/imgproc_cache.py \
