+++ _stbt/core.py
@@ -648,126 +648,6 @@ def argparser():
     return parser
 
 
-def _memoize_property_fn(fn):
-    @functools.wraps(fn)
-    def inner(self):
-        # pylint: disable=protected-access
-        if fn not in self._FrameObject__frame_object_cache:
-            self._FrameObject__frame_object_cache[fn] = fn(self)
-        return self._FrameObject__frame_object_cache[fn]
-    return inner
-
-
-def _mark_in_is_visible(fn):
-    @functools.wraps(fn)
-    def inner(self):
-        # pylint: disable=protected-access
-        try:
-            self._FrameObject__local.in_is_visible += 1
-        except AttributeError:
-            self._FrameObject__local.in_is_visible = 1
-        try:
-            return bool(fn(self))
-        finally:
-            self._FrameObject__local.in_is_visible -= 1
-    return inner
-
-
-def _noneify_property_fn(fn):
-    @functools.wraps(fn)
-    def inner(self):
-        # pylint: disable=protected-access
-        if (getattr(self._FrameObject__local, "in_is_visible", 0) or
-                self.is_visible):
-            return fn(self)
-        else:
-            return None
-    return inner
-
-
-class _FrameObjectMeta(type):
-    def __new__(mcs, name, parents, dct):
-        for k, v in dct.iteritems():
-            if isinstance(v, property):
-                # Properties must not have setters
-                if v.fset is not None:
-                    raise Exception(
-                        "FrameObjects must be immutable but this property has "
-                        "a setter")
-                f = v.fget
-                # The value of any property is cached after the first use
-                f = _memoize_property_fn(f)
-                # Public properties return `None` if the FrameObject isn't
-                # visible.
-                if k == 'is_visible':
-                    f = _mark_in_is_visible(f)
-                elif not k.startswith('_'):
-                    f = _noneify_property_fn(f)
-                dct[k] = property(f)
-
-        if 'AUTO_SELFTEST_EXPRESSIONS' not in dct:
-            dct['AUTO_SELFTEST_EXPRESSIONS'] = ['%s(frame={frame})' % name]
-
-        return super(_FrameObjectMeta, mcs).__new__(mcs, name, parents, dct)
-
-    def __init__(cls, name, parents, dct):
-        property_names = sorted([
-            p for p in dir(cls)
-            if isinstance(getattr(cls, p), property)])
-        assert 'is_visible' in property_names
-        cls._FrameObject__attrs = ["is_visible"] + sorted(
-            x for x in property_names
-            if x != "is_visible" and not x.startswith('_'))
-        super(_FrameObjectMeta, cls).__init__(name, parents, dct)
-
-
-class FrameObject(object):
-    __metaclass__ = _FrameObjectMeta
-
-    def __init__(self, frame):
-        if frame is None:
-            raise ValueError("FrameObject: frame must not be None")
-        self.__frame_object_cache = {}
-        self.__local = threading.local()
-        self._frame = frame
-
-    def __repr__(self):
-        args = ", ".join(("%s=%r" % x) for x in self._iter_attrs())
-        return "%s(%s)" % (self.__class__.__name__, args)
-
-    def _iter_attrs(self):
-        if self:
-            # pylint: disable=protected-access,no-member
-            for x in self.__class__.__attrs:
-                yield x, getattr(self, x)
-        else:
-            yield "is_visible", False
-
-    def __nonzero__(self):
-        return bool(self.is_visible)
-
-    def __cmp__(self, other):
-        # pylint: disable=protected-access
-        from itertools import izip_longest
-        if isinstance(other, self.__class__):
-            for s, o in izip_longest(self._iter_attrs(), other._iter_attrs()):
-                v = cmp(s[1], o[1])
-                if v != 0:
-                    return v
-            return 0
-        else:
-            return NotImplemented
-
-    def __hash__(self):
-        return hash(tuple(v for _, v in self._iter_attrs()))
-
-    @property
-    def is_visible(self):
-        raise NotImplementedError(
-            "Objects deriving from FrameObject must define an is_visible "
-            "property")
-
-
 # Internal
 # ===========================================================================
